+++ packages/babel-preset-metro-react-native/test/__fixtures__/App.ts
@@ -22,3 +22,5 @@ function takeDirection(direction: Direction): string {
 }
 
 console.log(takeDirection(Direction.Up));
+
+export { takeDirection };